pyproject.toml
[tool.poetry]
name = "loan-performance-dash"
version = "0.1.0"
description = "A Plotly Dash app for visualizing loan performance data for CAS deal."
authors = ["brandon-setegn <bjsetegn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.16.0"
pandas = "^2.2.1"
dash-bootstrap-components = "^1.5.0"
google-cloud-bigquery = "^3.17.2"
python-json-logger = "^2.0.7"
gunicorn = "^21.2.0"
db-dtypes = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
